requirements.txt
groq
python-dotenv
streamlit
discord.py
google-genai
gTTS